pyproject.toml
[project]
name = "qtf-mcp"
version = "0.2.0"
description = "A股数据 MCP 服务，支持多种数据源"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.22",
    "ta-lib>=0.6.8",
    "akshare>=1.17.94",
    "numpy>=1.26.0",
    "click>=8.3.1",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
qtf-mcp = "main:main"
